codie/style_learner/requirements.txt
fastapi
uvicorn
tree-sitter-languages
GitPython